feature: Fix Long to Int type mismatch after id type migration

+++ feature/recruitment/src/main/java/in/koreatech/koin/feature/recruitment/navigation/RecruitmentNavType.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 sealed class RecruitmentNavType {
     @Serializable
-    data class ApplicantManagement(val postId: Long) : RecruitmentNavType()
+    data class ApplicantManagement(val postId: Int) : RecruitmentNavType()
 
     @Serializable
     data object MyRecruitment : RecruitmentNavType()
